Update selection with new features
Newly created features of now added to the selection and previously selected features with only one part are deselected allowing the user to easily see where changes were made.
Note that the layer needs to be reloaded in order to update the Attribute Table. This can cause a significant delay with several thousands features if the table is opened.